Model Building and Execution

NEMStorageUnderUncertainty._build_storage_model — Method

_build_storage_model(
    storage::NEMStorageUnderUncertainty.StorageDevice,
    prices::Vector{<:AbstractFloat},
    times::Vector{Dates.DateTime},
    τ::Float64,
    ::NEMStorageUnderUncertainty.StandardArbitrage,
    ::NEMStorageUnderUncertainty.NoDegradation;
    silent,
    time_limit_sec,
    string_names
) -> JuMP.Model

Summary

Build a StandardArbitrage model

Arguments

storage: StorageDevice

prices: Energy prices in /MW/hr that corresponds to prices at times

times: Times to run model for

τ: Interval duration in hours

StandardArbitrage

silent: default false. true to suppress solver output

time_limit_sec: default nothing. Float64 to impose solver time limit in seconds

string_names: default true. false to disable JuMP string names

Returns

Built JuMP model

Methods

_build_storage_model(
    storage,
    prices,
    times,
    τ,
    ,
    ;
    silent,
    time_limit_sec,
    string_names
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/build_and_run.jl:23.

NEMStorageUnderUncertainty._build_storage_model — Method

_build_storage_model(
    storage::NEMStorageUnderUncertainty.StorageDevice,
    prices::Vector{<:AbstractFloat},
    times::Vector{Dates.DateTime},
    τ::Float64,
    throughput_limit::NEMStorageUnderUncertainty.StandardArbitrageThroughputLimit,
    ::NEMStorageUnderUncertainty.NoDegradation;
    silent,
    time_limit_sec,
    string_names
) -> JuMP.Model

Summary

Build a StandardArbitrageThroughputLimit model

Arguments

storage: StorageDevice

prices: Energy prices in /MW/hr that corresponds to prices at times

times: Times to run model for

τ: Interval duration in hours

StandardArbitrageThroughputLimit: StandardArbitrageThroughputLimit

silent: default false. true to suppress solver output

time_limit_sec: default nothing. Float64 to impose solver time limit in seconds

string_names: default true. false to disable JuMP string names

Returns

Built JuMP model

Methods

_build_storage_model(
    storage,
    prices,
    times,
    τ,
    throughput_limit,
    ;
    silent,
    time_limit_sec,
    string_names
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/build_and_run.jl:76.

NEMStorageUnderUncertainty.run_model — Method

run_model(
    optimizer::MathOptInterface.OptimizerWithAttributes,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    prices::Vector{<:AbstractFloat},
    times::Vector{Dates.DateTime},
    τ::Float64,
    formulation::NEMStorageUnderUncertainty.StorageModelFormulation,
    degradation::NEMStorageUnderUncertainty.DegradationModel;
    silent,
    time_limit_sec,
    string_names
) -> JuMP.Model

Summary

Runs a model using data in prices, times and τ (interval duration in hours). The type of model constructed and run is dependent on the formulation

Arguments

optimizer: A solver optimizer

storage: StorageDevice

prices: Energy prices in /MW/hr that corresponds to prices at times

times: Times to run model for

τ: Interval duration in hours

formulation: A model formulation (StorageModelFormulation)

Returns

A JuMP model if the solution is optimal (within solver tolerances)

A JuMP model with warning if a time/iteration limit is hit

Throws and error if infeasible/unbounded/etc.

Methods

run_model(
    optimizer,
    storage,
    prices,
    times,
    τ,
    formulation,
    degradation;
    silent,
    time_limit_sec,
    string_names
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/build_and_run.jl:133.