Price Data Compilers

NEMStorageUnderUncertainty.ActualData — Type

struct ActualData{T<:AbstractFloat} <: NEMStorageUnderUncertainty.NEMData

times::Vector{Dates.DateTime}

prices::Vector{T} where T<:AbstractFloat

τ::AbstractFloat

NEMStorageUnderUncertainty.ForecastData — Type

struct ForecastData{T<:AbstractFloat} <: NEMStorageUnderUncertainty.NEMData

run_times::Vector{Dates.DateTime}

forecasted_times::Vector{Dates.DateTime}

prices::Vector{T} where T<:AbstractFloat

τ::AbstractFloat

NEMStorageUnderUncertainty._get_data_by_times — Method

_get_data_by_times(
    forecast_data::DataFrames.DataFrame;
    forecasted_times,
    run_times
) -> DataFrames.DataFrame

Summary

Filters forecast prices based on supplied run_times (start, end) and forecasted_times (start, end)

Arguments

prices: ForecastPrice produced by get_all_forecast_prices

forecasted_times: (start_time, end_time), inclusive

run_times: (start_time, end_time), inclusive

Returns

Filtered ForecastPrice

Methods

_get_data_by_times(
    forecast_data;
    forecasted_times,
    run_times
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:189.

NEMStorageUnderUncertainty.get_all_actual_data — Method

get_all_actual_data(path::String) -> DataFrames.DataFrame

Summary

Obtains actual data from parquet files located at path

Arguments

path: Path to parquet partitions

Returns

DataFrame with settlement date, region and corresponding energy prices

Methods

get_all_actual_data(path)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:33.

NEMStorageUnderUncertainty.get_all_pd_and_p5_data — Method

get_all_pd_and_p5_data(
    pd_path::String,
    p5_path::String
) -> Tuple{DataFrames.DataFrame, DataFrames.DataFrame}

Summary

Obtains and compiles all forecasted price data from parquet files located at the P5MIN path (p5_path) and the PREDISPATCH path (pd_path)

Note that Parquet.jl cannot parse Timestamps from .parquet, so we use unix2datetime.

Arguments

pd_path: Path to PREDISPATCH parquet partitions

p5_path: Path to P5MIN parquet partitions

Returns

Compiled forecast data, with PREDISPATCH forecasts that overlap with P5MIN removed. Compiled forecast data has actual run times, forecasted times, regions and their corresponding energy prices.

Methods

get_all_pd_and_p5_data(pd_path, p5_path)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:92.