Objectives

NEMStorageUnderUncertainty._add_objective_standard! — Method

_add_objective_standard!(
    model::JuMP.Model,
    prices::Vector{<:AbstractFloat},
    times::Vector{Dates.DateTime},
    τ::Float64
) -> Any

Summary

Adds a standard revenue-maximising objective function:

\[\begin{aligned} \max_{t} \quad & \sum_{t}{\tau\lambda_t(p_t - q_t)} \end{aligned}\]

Arguments

model: JuMP model

prices: A Vector of prices in /MWh

times: A Vector of DateTimes

τ: Frequency of prices in hours

Methods

_add_objective_standard!(model, prices, times, τ)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/objectives.jl:17.

NEMStorageUnderUncertainty._add_objective_throughput_penalty! — Method

_add_objective_throughput_penalty!(
    model::JuMP.Model,
    prices::Vector{<:AbstractFloat},
    times::Vector{Dates.DateTime},
    τ::Float64,
    d_0::Float64,
    d_lifetime::Float64,
    e_rated::Float64,
    c_capital::Float64
) -> Any

Summary

Adds a revenue-maximising objective function that penalises storage throughput/cycling.

The penalty is the proportion of the warrantied throughput lifetime of the storage device expended during the modelled period, multiplied by the cost of a new storage device. In other words, the storage device replacement cost is amortised across throughput.

This bears similarities to the energy throughput model in this paper.

\[\begin{aligned} \max_{t} \quad & \sum_{t}^{T}{\tau\lambda_t(p_t - q_t)} - \frac{d_T - d_{0}}{d_{\textrm{lifetime}}}e_{\textrm{rated}}c_{\textrm{capital}} \end{aligned}\]

Arguments

model: JuMP model

prices: A Vector of prices in /MWh

times: A Vector of DateTimes

τ: Frequency of prices in hours

d_0: Initial throughput of storage device in MWh

d_lifetime: Warrantied throughput lifetime of the storage device in MWh

e_rated: Storage energy capacity in MWh

c_capital: Capital cost of storage device in AUD/MWh

Methods

_add_objective_throughput_penalty!(
    model,
    prices,
    times,
    τ,
    d_0,
    d_lifetime,
    e_rated,
    c_capital
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/objectives.jl:59.