Simulation Results Processing and Calculations

NEMStorageUnderUncertainty.calculate_summaries_and_vpl_vpi_across_scenarios — Method

calculate_summaries_and_vpl_vpi_across_scenarios(
    sim_folder::String
)

Summary

Summarises results, revenues and VPL and VPI for each simulated formulation

For each state, this function cycles through each simulated formulation and:

Calculates summary results (i.e. annual net revenue and mean relative gap)

Calculates revenues (i.e. annual net revenue, annual negative revenue)

Calculates the value of perfect lookahead and value of perfect information

A JLD2 file for each of these (with data for each simulated formulation) is released in the results folder

Arguments

sim_folder: Path containing simulations of different formulations and their results.

Returns

Nothing

lookahead.

Methods

calculate_summaries_and_vpl_vpi_across_scenarios(sim_folder)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/results.jl:257.

NEMStorageUnderUncertainty.calculate_vpl_vpi — Method

calculate_vpl_vpi(
    df::DataFrames.DataFrame
) -> DataFrames.DataFrame

Summary

Calculates values of perfect lookahead and information (as absolute values in AUD and as a percentage of perfect foresight revenue), and the detrimental decision metric

Value of perfect lookahead: What is the additional benefit (revenue) that a participant could gain if they were to know exactly what the market prices will be in the lookahead horizon.

$VPL = \textrm{Revenue}_\textrm{Actual Data Simulation} - \textrm{Revenue}_\textrm{Forecast Data Simulation}$

Value of perfect information: What is the additional benefit (revenue) that a participant could gain if they were to know exactly what the market prices will be over the entire year

$VPI = \textrm{Revenue}_\textrm{Perfect Foresight} - \textrm{Revenue}_\textrm{Forecast Data Simulation}$

Detrimental decision metric: What is the additional negative revenue (i.e. losses) incurred when using forecast market prices in the lookahead horizon as a percentage?

$DDM = \frac{\textrm{NegRev}_{\textrm{Actual Data Simulation}} -\textrm{NegRev}_{\textrm{Forecast Data Simulation}}}{\textrm{Revenue}_{\textrm{Actual Data Simulation}}-\textrm{Revenue}_{\textrm{Forecast Data Simulation}}}$

N.B. This function assumes that the input df only has data that corresponds to a device of a particular energy_capacity.

Arguments

df: DataFrame produced by _summarise_simulations

Returns

DataFrame with absolute values of perfect lookahead and information, and the same values as a percentage of perfect foresight revenue.

Methods

calculate_vpl_vpi(df)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/results.jl:137.

NEMStorageUnderUncertainty.calculate_vpl_vpi_across_scenarios — Method

calculate_vpl_vpi_across_scenarios(summary_folder::String)

Summary

Calculates values of perfect lookahead and information

For each state, this function cycles through each simulated formulation calculates the value of perfect lookahead and value of information

For a single state, the VPLs and VPIs across simulated formulations are then released in a JLD2 file in the results folder

Arguments

sim_folder: Path containing simulations of different formulations and their results.

Returns

Nothing

Methods

calculate_vpl_vpi_across_scenarios(summary_folder)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/results.jl:204.