Variables

NEMStorageUnderUncertainty._add_variable_charge_state! — Method

_add_variable_charge_state!(
    model::JuMP.Model,
    times::Vector{Dates.DateTime}
) -> JuMP.Containers.DenseAxisArray{JuMP.VariableRef, 1, Tuple{Vector{Dates.DateTime}}, Tuple{JuMP.Containers._AxisLookup{Dict{Dates.DateTime, Int64}}}}

Summary

Adds binary variable that indicates charging (i.e. $u_t=1$) when charging.

Arguments

model: JuMP model

times: A Vector of DateTimes

Methods

_add_variable_charge_state!(model, times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/variables.jl:55.

NEMStorageUnderUncertainty._add_variable_soc! — Method

_add_variable_soc!(
    model::JuMP.Model,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    times::Vector{Dates.DateTime}
) -> JuMP.Containers.DenseAxisArray

Summary

Adds variable that tracks state-of-charge (SoC, $e_t$).

The following variable bound is applied: $\underline{e} \leq e_t \leq \bar{e}$, where the limits represent the lower and upper SoC limits obtained from storage.

Arguments

model: JuMP model

storage: A StorageDevice

times: A Vector of DateTimes

Methods

_add_variable_soc!(model, storage, times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/variables.jl:39.

NEMStorageUnderUncertainty._add_variable_throughput! — Method

_add_variable_throughput!(
    model::JuMP.Model,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    times::Vector{Dates.DateTime}
) -> JuMP.Containers.DenseAxisArray

Summary

Adds variable that tracks throughput in MWh ($d_t$).

The following variable bound is applied: $d_0 \leq d_t$, where the limit represents the initial throughput obtained from storage.

Arguments

model: JuMP model

storage: A StorageDevice

times: A Vector of DateTimes

Methods

_add_variable_throughput!(model, storage, times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/variables.jl:72.

NEMStorageUnderUncertainty._add_variables_power! — Method

_add_variables_power!(
    model::JuMP.Model,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    times::Vector{Dates.DateTime}
) -> Tuple{JuMP.Containers.DenseAxisArray, JuMP.Containers.DenseAxisArray}

Summary

Adds variables for charging in MW ($q_t$) and discharging in MW ($p_t$).

The following variable bounds are applied:

$0 \leq p_t \leq \bar{p}$

$0 \leq q_t \leq \bar{p}$

Arguments

model: JuMP model

storage: A StorageDevice

times: A Vector of DateTimes

Methods

_add_variables_power!(model, storage, times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/model/variables.jl:14.