Price Data Compilers

NEMStorageUnderUncertainty.ActualPrice — Type

struct ActualPrice <: NEMStorageUnderUncertainty.PriceData

start_time::Dates.DateTime

end_time::Dates.DateTime

data::DataFrames.DataFrame

NEMStorageUnderUncertainty.ForecastPrice — Type

struct ForecastPrice <: NEMStorageUnderUncertainty.PriceData

run_start::Dates.DateTime

run_end::Dates.DateTime

forecasted_start::Dates.DateTime

forecasted_end::Dates.DateTime

data::DataFrames.DataFrame

NEMStorageUnderUncertainty._get_prices_by_times — Method

_get_prices_by_times(
    prices::NEMStorageUnderUncertainty.ActualPrice,
    times::Union{Nothing, Tuple{Dates.DateTime, Dates.DateTime}}
) -> NEMStorageUnderUncertainty.ActualPrice

Summary

Filters actual prices based on supplied times (start, end)

Arguments

prices: ActualPrice produced by get_all_actual_prices

times: (start_time, end_time), inclusive

Returns

Filtered ActualPrice

Methods

_get_prices_by_times(prices, times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:175.

NEMStorageUnderUncertainty._get_prices_by_times — Method

_get_prices_by_times(
    prices::NEMStorageUnderUncertainty.ForecastPrice;
    forecasted_times,
    run_times
) -> NEMStorageUnderUncertainty.ForecastPrice

Summary

Filters forecast prices based on supplied run_times (start, end) and forecasted_times (start, end)

Arguments

prices: ForecastPrice produced by get_all_forecast_prices

forecasted_times: (start_time, end_time), inclusive

run_times: (start_time, end_time), inclusive

Returns

Filtered ForecastPrice

Methods

_get_prices_by_times(prices; forecasted_times, run_times)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:126.

NEMStorageUnderUncertainty.get_all_actual_prices — Method

get_all_actual_prices(
    path::String
) -> NEMStorageUnderUncertainty.ActualPrice

Summary

Obtains actual price data from parquet files located at path

Arguments

path: Path to parquet partitions

Returns

DataFrame with settlement date, region and corresponding energy prices

Methods

get_all_actual_prices(path)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:32.

NEMStorageUnderUncertainty.get_all_forecast_prices — Method

get_all_forecast_prices(
    pd_path::String,
    p5_path::String
) -> NEMStorageUnderUncertainty.ForecastPrice

Summary

Obtains and compiles all forecasted price data from parquet files located at the P5MIN path (p5_path) and the PREDISPATCH path (pd_path)

Note that Parquet.jl cannot parse Timestamps from .parquet, so we use unix2datetime.

Arguments

pd_path: Path to PREDISPATCH parquet partitions

p5_path: Path to P5MIN parquet partitions

Returns

Compiled forecast data, with PREDISPATCH forecasts that overlap with P5MIN removed. Compiled forecast data has actual run times, forecasted times, regions and their corresponding energy prices.

Methods

get_all_forecast_prices(pd_path, p5_path)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/data.jl:64.