Price Data Compilers

NEMStorageUnderUncertainty.get_all_actual_prices — Method

get_all_actual_prices(path::String)

Obtains actual price data from parquet files located at path

Arguments

path: Path to parquet partitions

Returns

DataFrame with settlement date, region and corresponding energy prices

NEMStorageUnderUncertainty.get_all_forecast_prices — Method

get_all_forecast_prices(pd_path::String, p5_path::String)

Obtains and compiles all forecasted price data from parquet files located at the P5MIN path (p5_path) and the PREDISPATCH path (pd_path)

Note that Parquet.jl cannot parse Timestamps from .parquet, so we use unix2datetime.

Arguments

pd_path: Path to PREDISPATCH parquet partitions

p5_path: Path to P5MIN parquet partitions

Returns

Compiled forecast data, with PREDISPATCH forecasts that overlap with P5MIN removed. Compiled forecast data has actual run times, forecasted times, regions and their corresponding energy prices.

NEMStorageUnderUncertainty.get_forecast_prices_by_times — Method

get_forecast_prices_by_times(
forecast_prices::DataFrame,
forecasted_times::Tuple{DateTime, DateTime}=nothing,
run_times::Tuple{DateTime, DateTime}=nothing
)

Filters forecast prices based on supplied run_times (start, end) and forecasted_times (start, end)

Arguments

forecast_prices: DataFrame produced by get_all_forecast_prices

forecasted_times: (starttime, endtime), inclusive

run_times: (starttime, endtime), inclusive

Returns

Filtered forecast prices