Simulation Results Processing and Calculations

NEMStorageUnderUncertainty.calculate_summaries_and_vpi_vpf_across_scenarios — Method

calculate_summaries_and_vpi_vpf_across_scenarios(
    sim_folder::String
) -> Dict{String, DataFrames.DataFrame}

Summary

Summarises results for each simulated formulation and then calculates values of perfect information and foresight.

For each state, this function cycles through each simulated formulation and:

Calculates summary results (i.e. annual revenue and mean relative gap)

Calculates the value of perfect information and value of perfect foresight

For a single state, the VPIs and VPFs across simulated formulations are then released information a JLD2 file in the results folder, along with summary results for each simulated formulation.

Arguments

sim_folder: Path containing simulations of different formulations and their results.

Returns

Dict mapping each state to DataFrame with VPI and VPF for each formulation and lookahead.

Methods

calculate_summaries_and_vpi_vpf_across_scenarios(sim_folder)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/results.jl:192.

NEMStorageUnderUncertainty.calculate_vpi_vpf — Method

calculate_vpi_vpf(
    df::DataFrames.DataFrame
) -> DataFrames.DataFrame

Summary

Calculates values of perfect information and foresight as absolute values (in AUD) and as a percentage of perfect foresight revenue.

Value of perfect information: What is the additional benefit (revenue) that a participant could gain if they were to know exactly what the market prices will be in the lookahead horizon.

$VPI = \textrm{Revenue}_\textrm{Actual Data Simulation} - \textrm{Revenue}_\textrm{Forecast Data Simulation}$

Value of perfect foresight: What is the additional benefit (revenue) that a participant could gain if they were to know exactly what the market prices will be over the entire year

$VPF = \textrm{Revenue}_\textrm{Perfect Foresight} - \textrm{Revenue}_\textrm{Forecast Data Simulation}$

N.B. This function assumes that the input df only has data that corresponds to a device of a particular energy_capacity.

Arguments

df: DataFrame produced by _summarise_simulations

Returns

DataFrame with absolute values of perfect information and foresight, and the same values as a percentage of perfect foresight revenue.

Methods

calculate_vpi_vpf(df)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/results.jl:128.