Simulation Functions

NEMStorageUnderUncertainty._get_periods_for_simulation — Method

_get_periods_for_simulation(
    decision_start_time::Dates.DateTime,
    decision_end_time::Dates.DateTime,
    binding::Dates.Period,
    horizon::Dates.Period,
    data::NEMStorageUnderUncertainty.ActualData
) -> DataFrames.DataFrame

Summary

Gets decision points, binding intervals and horizon ends given ActualData and simulation parameters.

Arguments

decision_start_time: Decision start time.

decision_end_time: Decision end time.

binding: decision_time + binding gives the last binding period

horizon: decision_time + horizon gives the end of the simulation horizon

data: ActualData

Returns

DataFrame with the following columns:

decision_interval: data.times indices that correspond to decision points

binding_start: data.times indices that correspond to the first binding period for each simulation.

binding_end: data.times indices that correspond to the last binding period for each simulation.

horizon_end: data.times indices that correspond to the horizon end for each simulation

Methods

_get_periods_for_simulation(
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    data
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:26.

NEMStorageUnderUncertainty._get_periods_for_simulation — Method

_get_periods_for_simulation(
    decision_start_time::Dates.DateTime,
    decision_end_time::Dates.DateTime,
    binding::Dates.Period,
    horizon::Dates.Period,
    data::NEMStorageUnderUncertainty.ForecastData
) -> DataFrames.DataFrame

Summary

Gets decision points, binding intervals and horizon ends given ForecastData and simulation parameters.

Arguments

decision_start_time: Decision start time. Applies to run_time

decision_end_time: Decision end time. Applies to run_time

binding: decision_time + binding gives the last binding period. Applies to forecasted_time

horizon: decision_time + horizon gives the end of the simulation horizon. Applies to forecasted_time

data: ForecastData

Returns

DataFrame with the following columns:

decision_interval: data.times indices that correspond to run_time decision points

binding_start: data.times indices that correspond to the first binding period for each simulation. Applies to forecasted_time.

binding_end: data.times indices that correspond to the last binding period for each simulation. Applies to forecasted_time

horizon_end: data.times indices that correspond to the horizon end for each simulation. Applies to forecasted_time

Methods

_get_periods_for_simulation(
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    data
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:116.

NEMStorageUnderUncertainty._update_storage_state — Method

_update_storage_state(
    storage::NEMStorageUnderUncertainty.StorageDevice,
    binding_results::DataFrames.DataFrame,
    _::NEMStorageUnderUncertainty.NoDegradation
) -> Any

Summary

"Updates" (via new StorageDevice) storage state between model runs. Specifically:

Updates soc₀ to reflect soc at end of binding decisions from last model run

Updates storage throughput based on model run

throughput is defined as discharged energy, and hence does not consider η_discharge

Arguments

storage: StorageDevice

binding_results: Binding results from last model run

τ: Interval duration in hours

degradation: No degradation model NoDegradation

Returns

New StorageDevice with updated soc₀ and throughput

Methods

_update_storage_state(storage, binding_results, _)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:258.

NEMStorageUnderUncertainty.simulate_storage_operation — Method

simulate_storage_operation(
    optimizer::MathOptInterface.OptimizerWithAttributes,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    data::NEMStorageUnderUncertainty.ActualData,
    model_formulation::NEMStorageUnderUncertainty.StorageModelFormulation,
    degradation::NEMStorageUnderUncertainty.DegradationModel;
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    capture_all_decisions,
    silent,
    show_progress,
    time_limit_sec,
    string_names,
    relative_gap_in_results
)

Summary

Simulate storage operation using actual data

Arguments

optimizer: A solver optimizer

storage: StorageDevice

data: ActualData

model_formulation: A model formulation (StorageModelFormulation)

degradation: A degradation model (DegradationModel)

decision_start_time: Decision start time

decision_end_time: Decision end time

binding: Binding horizon, out from decision_start_time

horizon: Model lookahead horizon, out from decision_start_time

capture_all_decisions: Default false. If true, report non binding decisions in results

silent: default false. true to suppress solver output

time_limit_sec: default nothing. Float64 to impose solver time limit in seconds

string_names: default true. false to disable JuMP string names

relative_gap_in_results: default false. If true, adds a column with relative MIP gap in results.

Returns

Simulation results

Methods

simulate_storage_operation(
    optimizer,
    storage,
    data,
    model_formulation,
    degradation;
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    capture_all_decisions,
    silent,
    show_progress,
    time_limit_sec,
    string_names,
    relative_gap_in_results
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:294.

NEMStorageUnderUncertainty.simulate_storage_operation — Method

simulate_storage_operation(
    optimizer::MathOptInterface.OptimizerWithAttributes,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    data::NEMStorageUnderUncertainty.ForecastData,
    model_formulation::NEMStorageUnderUncertainty.StorageModelFormulation,
    degradation::NEMStorageUnderUncertainty.DegradationModel;
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    capture_all_decisions,
    silent,
    show_progress,
    time_limit_sec,
    string_names,
    relative_gap_in_results
)

Summary

Simulate storage operation using forecast data

Arguments

optimizer: A solver optimizer

storage: StorageDevice

data: ForecastData

model_formulation: A model formulation (StorageModelFormulation)

degradation: A degradation model (DegradationModel)

decision_start_time: Decision start time

decision_end_time: Decision end time

binding: Binding horizon, out from decision_start_time

horizon: Model lookahead horizon, out from decision_start_time

capture_all_decisions: Default false. If true, report non binding decisions in results

silent: default false. true to suppress solver output

time_limit_sec: default nothing. Float64 to impose solver time limit in seconds

string_names: default true. false to disable JuMP string names

relative_gap_in_results: default false. If true, adds a column with relative MIP gap in results.

Returns

Simulation results

Methods

simulate_storage_operation(
    optimizer,
    storage,
    data,
    model_formulation,
    degradation;
    decision_start_time,
    decision_end_time,
    binding,
    horizon,
    capture_all_decisions,
    silent,
    show_progress,
    time_limit_sec,
    string_names,
    relative_gap_in_results
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:405.