Simulations

NEMStorageUnderUncertainty._run_model — Method

_run_model(
    optimizer::DataType,
    storage::NEMStorageUnderUncertainty.StorageDevice,
    prices::Vector{<:Float64},
    times::Vector{Dates.DateTime},
    τ::Float64,
    formulation::NEMStorageUnderUncertainty.StorageModelFormulation
) -> JuMP.Model

Summary

Runs a model using data in prices, times and τ (interval duration in hours). The type of model constructed and run is dependent on the formulation

Arguments

optimizer: A solver optimizer

storage: StorageDevice

prices: Energy prices in /MW/hr that corresponds to prices at times

times: Times to run model for

τ: Interval duration in hours

formulation: A model formulation (StorageModelFormulation)

Returns

A JuMP model if the solution is optimal (within solver tolerances)

A JuMP model with warning if a time/iteration limit is hit

Throws and error if infeasible/unbounded/etc.

Methods

_run_model(
    optimizer,
    storage,
    prices,
    times,
    τ,
    formulation
)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:19.

NEMStorageUnderUncertainty._update_storage_state — Method

_update_storage_state(
    storage::NEMStorageUnderUncertainty.StorageDevice,
    model::JuMP.Model,
    τ::Float64,
    degradation::NEMStorageUnderUncertainty.NoDegradation
) -> Any

Summary

"Updates" (via new StorageDevice) storage state between model runs. Specifically:

Updates soc₀ to reflect soc at end of last model run

Considers η_discharge and η_charge (explicitly/via intertemporal SoC constraints)

Updates storage throughput based on model run

Based on discharged (delivered) energy and thus does not consider any η

Arguments

storage: StorageDevice

model: Model from _run_model with solution values

τ: Interval duration in hours

degradation: No degradation model NoDegradation

Returns

New StorageDevice with updated soc₀ and throughput

Methods

_update_storage_state(storage, model, τ, degradation)

defined at /home/runner/work/NEMStorageUnderUncertainty/NEMStorageUnderUncertainty/src/simulations.jl:64.